Refactor uploading to avoid code duplication
All the code after taking a screenshot was moved to the storage package.
Strictly speaking, the package does more than just storage now, but all
functionality is about uploading and what happens after that
(like notifications, clipboard), so it makes much more sense to put it
there.
